# contracts/openapi/kelimio-api.yaml
openapi: 3.1.0
info:
  title: Kelimio API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Versioned contract for the Kelimio mobile and administrative clients. Server
    state is authoritative for identity, answers, score, energy, progress, and
    entitlements.
...
servers:
  - url: /
    description: Base URL supplied by the environment-specific client configuration.
...
security:
  - bearerAuth: []
paths:
...
  /v1/tests/{testId}/attempts:
    post:
      operationId: startAttempt
      tags: [Learning]
      summary: Start an online attempt for the current test revision
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Id"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ClientCapabilities"
      responses:
        "201":
          description: Attempt created; answer keys are deliberately absent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResponse"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/attempts/{attemptId}/answers:
    post:
      operationId: submitAnswer
      tags: [Learning]
      summary: Record and evaluate one online answer exactly once
      description: >-
        Reusing submissionId returns the previously committed response and never
        creates a second attempt fact, score event, energy event, or outbox event.
        The complete request body is limited to 8192 bytes and is rejected before
        JSON allocation or transactional command handling when that limit is exceeded.
      x-kelimio-max-request-body-bytes: 8192
      parameters:
        - $ref: "#/components/parameters/AttemptId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ClientCapabilities"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerRequest"
      responses:
        "200":
          description: Authoritative evaluation and resulting server state
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerRecordedResponse"
        "401":
          $ref: "#/components/responses/Problem"
        "413":
          description: Request body exceeds the answer-submission byte limit
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/attempts/{attemptId}/answers/{submissionId}:
    get:
      operationId: getRecordedAnswer
      tags: [Learning]
      summary: Reconcile one previously committed answer owned by the current user
      description: >-
        Returns the immutable committed result only when both the attempt and
        submission belong to the authenticated user. Missing or non-owned
        records are indistinguishable and return not found.
      parameters:
        - $ref: "#/components/parameters/AttemptId"
        - name: submissionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Id"
        - $ref: "#/components/parameters/ClientCapabilities"
      responses:
        "200":
          description: Previously committed authoritative answer result
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerRecordedResponse"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/attempts/{attemptId}/finish:
    post:
      operationId: finishAttempt
      tags: [Learning]
      summary: Finish an attempt after all planned questions are answered
      parameters:
        - $ref: "#/components/parameters/AttemptId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ClientCapabilities"
      responses:
        "200":
          description: Final attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishAttemptResponse"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token obtained with Authorization Code and PKCE.
  headers:
...
    NoStore:
      description: Prevents sensitive answer feedback and owner-scoped import data from being cached.
      schema:
        type: string
        const: no-store
  parameters:
...
    AttemptId:
      name: attemptId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable UUID generated once for the logical command.
      schema:
        type: string
        format: uuid
    ClientCapabilities:
      name: X-Kelimio-Client-Capabilities
      in: header
      required: false
      description: >-
        Comma-separated bounded capability tokens implemented by the client.
        Missing means the empty set; this is a compatibility signal, not authorization.
      schema:
        type: string
        minLength: 1
        maxLength: 1040
        pattern: "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*)+(?:, ?[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*)+)*$"
  responses:
    Problem:
      description: RFC 9457 problem details
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Id:
      type: string
      format: uuid
    MatchingItemId:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      description: >-
        Independently generated random UUIDv4 for one public matching-side
        item. It is unrelated to text, authored pair identity, insertion
        order, the opposite-side item ID, or any shared namespace.
    QuestionPrompt:
      type: [string, "null"]
      description: >-
        Required question prompt value. MATCHING carries an explicit null;
        every other question type carries nonblank target-language text.
    LanguageTag:
      type: string
      description: >-
        Canonically cased BCP 47 subset: lowercase primary language, optional
        title-case script, uppercase region, and lowercase variants. Extensions
        and private-use subtags are outside the initial API contract.
      minLength: 2
      maxLength: 35
      pattern: "^[a-z]{2,8}(?:-[A-Z][a-z]{3})?(?:-(?:[A-Z]{2}|[0-9]{3}))?(?:-(?:[a-z0-9]{5,8}|[0-9][a-z0-9]{3}))*$"
      examples: [tr, en, ar, fr, pt-BR, zh-Hant]
...
    AttemptResponse:
      type: object
      additionalProperties: false
      required: [id, testId, testRevisionId, supportLanguage, state, questions, startedAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        testId:
          $ref: "#/components/schemas/Id"
        testRevisionId:
          $ref: "#/components/schemas/Id"
        supportLanguage:
          $ref: "#/components/schemas/LanguageTag"
        state:
          $ref: "#/components/schemas/AttemptState"
        questions:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/QuestionPayload"
        startedAt:
          type: string
          format: date-time
    QuestionPayload:
      type: object
      additionalProperties: false
      x-kelimio-test-nullable-matching-prompt: true
      description: >-
        Answer-key-free online question payload. MULTIPLE_CHOICE_CLOZE and
        TYPED_CLOZE carry a raw target-language sentence and exactly one literal
        ASCII --- marker; the client renders that marker as an accessible
        visual blank. MATCHING has a null prompt, no options, and independently
        ordered target/support item arrays with no field linking the two sides.
      x-kelimio-cross-field-invariants:
        - MULTIPLE_CHOICE_CLOZE and TYPED_CLOZE prompts contain exactly one literal ASCII ---; overlapping occurrences count.
        - WORD_MULTIPLE_CHOICE and MULTIPLE_CHOICE_CLOZE contain exactly four options; TYPED_CLOZE and MATCHING contain none.
        - MATCHING contains two through six target items and the same number of support items; all item IDs are unique per side and the two ID sets are disjoint.
        - Non-MATCHING questions contain no target or support items.
      required: [questionId, questionRevisionId, type, position, prompt, options, targetItems, supportItems]
      allOf:
        - if:
            properties:
              type:
                enum: [MULTIPLE_CHOICE_CLOZE, TYPED_CLOZE]
            required: [type]
          then:
            properties:
              prompt:
                pattern: '^(?![\s\S]*----)(?![\s\S]*---[\s\S]*---)[\s\S]*---[\s\S]*$'
        - if:
            properties:
              type:
                enum: [WORD_MULTIPLE_CHOICE, MULTIPLE_CHOICE_CLOZE]
            required: [type]
          then:
            properties:
              options:
                minItems: 4
                maxItems: 4
          else:
            properties:
              options:
                maxItems: 0
        - if:
            properties:
              type:
                const: MATCHING
            required: [type]
          then:
            allOf:
              - properties:
                  prompt:
                    type: "null"
              - oneOf:
                  - properties:
                      targetItems: { minItems: 2, maxItems: 2 }
                      supportItems: { minItems: 2, maxItems: 2 }
                  - properties:
                      targetItems: { minItems: 3, maxItems: 3 }
                      supportItems: { minItems: 3, maxItems: 3 }
                  - properties:
                      targetItems: { minItems: 4, maxItems: 4 }
                      supportItems: { minItems: 4, maxItems: 4 }
                  - properties:
                      targetItems: { minItems: 5, maxItems: 5 }
                      supportItems: { minItems: 5, maxItems: 5 }
                  - properties:
                      targetItems: { minItems: 6, maxItems: 6 }
                      supportItems: { minItems: 6, maxItems: 6 }
          else:
            properties:
              prompt:
                type: string
                minLength: 1
                maxLength: 1000
              targetItems:
                maxItems: 0
              supportItems:
                maxItems: 0
      properties:
        questionId:
          $ref: "#/components/schemas/Id"
        questionRevisionId:
          $ref: "#/components/schemas/Id"
        type:
          type: string
          enum: [WORD_MULTIPLE_CHOICE, MULTIPLE_CHOICE_CLOZE, TYPED_CLOZE, MATCHING]
        position:
          type: integer
          minimum: 1
        prompt:
          $ref: "#/components/schemas/QuestionPrompt"
          x-kelimio-required-nullable: true
        options:
          type: array
          items:
            $ref: "#/components/schemas/AnswerOption"
        targetItems:
          type: array
          items:
            $ref: "#/components/schemas/MatchingItem"
        supportItems:
          type: array
          items:
            $ref: "#/components/schemas/MatchingItem"
    AnswerOption:
      type: object
      additionalProperties: false
      required: [id, text]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        text:
          type: string
          minLength: 1
          maxLength: 500
    MatchingItem:
      type: object
      additionalProperties: false
      required: [id, text]
      properties:
        id:
          $ref: "#/components/schemas/MatchingItemId"
        text:
          type: string
          minLength: 1
          maxLength: 500
    MatchingSelection:
      type: object
      additionalProperties: false
      x-kelimio-redacted-to-string: true
      x-kelimio-redaction-test-matching-selection: true
      required: [targetItemId, supportItemId]
      properties:
        targetItemId:
          x-kelimio-sensitive: true
          allOf:
            - $ref: "#/components/schemas/MatchingItemId"
        supportItemId:
          x-kelimio-sensitive: true
          allOf:
            - $ref: "#/components/schemas/MatchingItemId"
    SubmitAnswerRequest:
      type: object
      additionalProperties: false
      description: >-
        Carries exactly one answer form. typedAnswer and matches are sensitive,
        are evaluated only by the server, and must never be logged or emitted
        to analytics or outbox events.
      x-kelimio-cross-field-invariants:
        - Exactly one of selectedOptionId, typedAnswer, and matches is present.
        - matches is a complete bijection over the server-issued matching item IDs, with no duplicate target or support ID.
      x-kelimio-redacted-to-string: true
      x-kelimio-redaction-test-submit-answer-request: true
      required: [submissionId, questionRevisionId]
      oneOf:
        - required: [selectedOptionId]
          not:
            anyOf:
              - required: [typedAnswer]
              - required: [matches]
        - required: [typedAnswer]
          not:
            anyOf:
              - required: [selectedOptionId]
              - required: [matches]
        - required: [matches]
          not:
            anyOf:
              - required: [selectedOptionId]
              - required: [typedAnswer]
      properties:
        submissionId:
          $ref: "#/components/schemas/Id"
        questionRevisionId:
          $ref: "#/components/schemas/Id"
        selectedOptionId:
          $ref: "#/components/schemas/Id"
        typedAnswer:
          type: string
          minLength: 1
          maxLength: 500
          writeOnly: true
          x-kelimio-sensitive: true
        matches:
          type: array
          minItems: 2
          maxItems: 6
          items:
            $ref: "#/components/schemas/MatchingSelection"
          writeOnly: true
          x-kelimio-sensitive: true
    AnswerRecordedResponse:
      type: object
      additionalProperties: false
      description: >-
        Post-commit feedback for exactly one submitted question. Multiple-choice
        responses contain correctOptionId; typed-cloze responses contain only
        the primary authored correctAnswerText; matching responses contain the
        complete correctMatches mapping. None appears before commit.
      x-kelimio-cross-field-invariants:
        - Exactly one of correctOptionId, correctAnswerText, and correctMatches is present.
        - correctMatches exactly covers the target and support item IDs issued for the submitted matching question.
      x-kelimio-redacted-to-string: true
      x-kelimio-redaction-test-answer-recorded-response: true
      required:
        - submissionId
        - correct
        - activeScoreDelta
        - lifetimeScoreDelta
        - activeQuestionScore
        - lifetimeScore
        - energy
        - attemptState
      oneOf:
        - required: [correctOptionId]
          not:
            anyOf:
              - required: [correctAnswerText]
              - required: [correctMatches]
        - required: [correctAnswerText]
          not:
            anyOf:
              - required: [correctOptionId]
              - required: [correctMatches]
        - required: [correctMatches]
          not:
            anyOf:
              - required: [correctOptionId]
              - required: [correctAnswerText]
      properties:
        submissionId:
          $ref: "#/components/schemas/Id"
        correct:
          type: boolean
        correctOptionId:
          x-kelimio-sensitive: true
          allOf:
            - $ref: "#/components/schemas/Id"
        correctAnswerText:
          type: string
          minLength: 1
          maxLength: 500
          x-kelimio-sensitive: true
        correctMatches:
          type: array
          minItems: 2
          maxItems: 6
          items:
            $ref: "#/components/schemas/MatchingSelection"
          x-kelimio-sensitive: true
        activeScoreDelta:
          type: integer
          minimum: 0
          maximum: 60
        lifetimeScoreDelta:
          type: integer
          minimum: 0
          maximum: 60
        activeQuestionScore:
          type: integer
          minimum: 0
          maximum: 60
        lifetimeScore:
          type: integer
          minimum: 0
        energy:
          $ref: "#/components/schemas/EnergyResponse"
        attemptState:
          $ref: "#/components/schemas/AttemptState"
    FinishAttemptResponse:
      type: object
      additionalProperties: false
      required: [attemptId, state, correctCount, questionCount, correctRatio, completedAt]
      properties:
        attemptId:
          $ref: "#/components/schemas/Id"
        state:
          $ref: "#/components/schemas/AttemptState"
        correctCount:
          type: integer
          minimum: 0
        questionCount:
          type: integer
          minimum: 1
        correctRatio:
          type: number
          minimum: 0
          maximum: 1
        completedAt:
          type: string
          format: date-time
    AttemptState:
      type: string
      enum: [IN_PROGRESS, COMPLETED_PASS, COMPLETED_FAIL, INTERRUPTED_ENERGY]
    EnergyResponse:
      type: object
      additionalProperties: false
      required: [balance, maximum, unlimited, asOf]
      properties:
        balance:
          type: integer
          minimum: 0
          maximum: 5
        maximum:
          type: integer
          const: 5
        unlimited:
          type: boolean
        nextRegenerationAt:
          type: [string, "null"]
          format: date-time
        asOf:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        requestId:
          type: string